refactor: move the socket event layer into useSocketEvents (Q10, F4)
App's 414-line effect becomes a hook that owns the listing counter,
batch-reindex tracker and last-downloaded ref, registering its eighteen
listeners from one table so the parallel .off block cannot drift.
1,053 lines down to 588.
